Remove category=cited filter from list of author's publications
Line 25 of 10_bibliography.tex  ("\nocite{li2002design}") has no effect when \printbibliography contains category=cited. It is therefore removed.

Note that this gives rise to a further issue: The non-cited references do not have a citation number/name given but still are prefixed by the labelprefix. Both issues probably never appear in a real thesis but are startling during writing.